+++ global_helpers/panther_gsuite_helpers.yml
@@ -1,0 +1,5 @@
+AnalysisType: global
+Filename: panther_gsuite_helpers.py
+GlobalID: "panther_gsuite_helpers"
+Description: >
+  Used to define global helpers and variables for GSuite events.
