+++ rules/gsuite_reports_rules/gsuite_drive_overly_visible.py
@@ -1,5 +1,5 @@
-from panther_base_helpers import gsuite_details_lookup as details_lookup
-from panther_base_helpers import gsuite_parameter_lookup as param_lookup
+from panther_gsuite_helpers import gsuite_details_lookup as details_lookup
+from panther_gsuite_helpers import gsuite_parameter_lookup as param_lookup
 
 RESOURCE_CHANGE_EVENTS = {
     "create",
