+++ rules/gsuite_reports_rules/gsuite_drive_visibility_change.py
@@ -1,7 +1,7 @@
 import json
 from unittest.mock import MagicMock
 
-from panther_base_helpers import gsuite_parameter_lookup as param_lookup
+from panther_gsuite_helpers import gsuite_parameter_lookup as param_lookup
 
 # Add any domain name(s) that you expect to share documents with in the ALLOWED_DOMAINS set
 ALLOWED_DOMAINS = set()
